+++ src/Schema/Factories/FieldFactory.php
@@ -68,13 +68,14 @@ class FieldFactory
             $fieldValue = $fieldValue->useDefaultResolver();
         }
 
-        $fieldMiddleware = $this->directiveFactory->associatedOfType($fieldDefinitionNode, FieldMiddleware::class)
-            // Middleware resolve in reversed order
-            ->push(app(RenameArgsDirective::class))
-            ->push(app(SpreadDirective::class))
-            ->push(app(TransformArgsDirective::class))
-            ->push(app(ValidateDirective::class))
-            ->push(app(SanitizeDirective::class));
+        $fieldMiddleware = $this->directiveFactory->associatedOfType($fieldDefinitionNode, FieldMiddleware::class);
+
+        foreach (config('lighthouse.field_middleware') as $globalFieldMiddlewareClass) {
+            // Middleware resolve in reversed order, so we add them to the beginning of the chain
+            $fieldMiddleware->prepend(
+                app($globalFieldMiddlewareClass)
+            );
+        }
 
         $resolverWithMiddleware = $this->pipeline
             ->send($fieldValue)